# Fix hash collision vulnerability of `toMap` call and low speed of `as[T]/toMap/toSeq` calls