Default hash maps in Scala handle collisions using an algorithm with `O(n^2)` complexity.

In this PR I propose to switch API from `Map/Seq` to `collection.Map/collection.Seq` and get `O(n*log(n))` complexity for the handling of hash map collisions. Please see bellow results of [benchmarks](https://github.com/plokhotnyuk/jsoniter-scala/pull/352) for parsing of JSON objects of different sizes when all keys have 0 value of hash code.
There is an option of using `mutable.CollisionProofHashMap`, but it still require of switching the type of result of the `toMap` call and should be ported for Scala 2.11.x/2.12.x for that (see https://github.com/scala/scala-collection-compat/issues/234).

Another option that can work for immutable result of `toMap` would be limiting numbers of keys that can be converted by `toMap`.

Yet another problem with too expensive `typeOf` calls that was detected by these benchmarks and raised [here](https://github.com/pathikrit/dijon/issues/30) is fixed in this PR by removing usage of Scala reflection. So instead of `as[Int]/as[String]/...` the following methods were implemented: `asInt/asString/...`. Also, the `asNone` method is added, here is a possible use case for it:
```
scala> import com.github.pathikrit.dijon._
import com.github.pathikrit.dijon._

scala> val j = json"""{"x":["1",null,{}]}"""
j: com.github.pathikrit.dijon.package.Json[Any] = {"x":["1",null,{}]}

scala> j.x.toSeq.filterNot(_.asNone.isDefined)
res0: scala.collection.Seq[com.github.pathikrit.dijon.package.SomeJson] = ArrayBuffer("1", {})
```

## Before
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                                    (size)   Mode  Cnt        Score        Error  Units
[info] ExtractFieldsReading.dijonParse                   1  thrpt    3  2983289.709 ± 830261.728  ops/s
[info] ExtractFieldsReading.dijonParse                  10  thrpt    3   881482.792 ±  17872.802  ops/s
[info] ExtractFieldsReading.dijonParse                 100  thrpt    3    44245.410 ±   8167.595  ops/s
[info] ExtractFieldsReading.dijonParse                1000  thrpt    3     3884.390 ±     55.402  ops/s
[info] ExtractFieldsReading.dijonParse               10000  thrpt    3      295.359 ±     11.218  ops/s
[info] ExtractFieldsReading.dijonParse              100000  thrpt    3       23.642 ±     20.693  ops/s
[info] ExtractFieldsReading.dijonParseAndThenToMap       1  thrpt    3    65960.415 ±  34599.506  ops/s
[info] ExtractFieldsReading.dijonParseAndThenToMap      10  thrpt    3    54130.569 ±  27457.571  ops/s
[info] ExtractFieldsReading.dijonParseAndThenToMap     100  thrpt    3    10956.483 ±   5704.414  ops/s
[info] ExtractFieldsReading.dijonParseAndThenToMap    1000  thrpt    3      190.719 ±     14.853  ops/s
[info] ExtractFieldsReading.dijonParseAndThenToMap   10000  thrpt    3        1.784 ±      1.016  ops/s
[info] ExtractFieldsReading.dijonParseAndThenToMap  100000  thrpt    3        0.009 ±      0.001  ops/s
```
## After
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                                    (size)   Mode  Cnt        Score         Error  Units
[info] ExtractFieldsReading.dijonParse                   1  thrpt    3  6274412.043 ±  562232.549  ops/s
[info] ExtractFieldsReading.dijonParse                  10  thrpt    3   978256.767 ±  292028.701  ops/s
[info] ExtractFieldsReading.dijonParse                 100  thrpt    3    42670.207 ±     526.796  ops/s
[info] ExtractFieldsReading.dijonParse                1000  thrpt    3     3090.553 ±    2022.896  ops/s
[info] ExtractFieldsReading.dijonParse               10000  thrpt    3      249.881 ±      86.810  ops/s
[info] ExtractFieldsReading.dijonParse              100000  thrpt    3       18.353 ±      10.673  ops/s
[info] ExtractFieldsReading.dijonParseAndThenToMap       1  thrpt    3  6840144.250 ± 1362093.932  ops/s
[info] ExtractFieldsReading.dijonParseAndThenToMap      10  thrpt    3   982408.460 ±  467516.116  ops/s
[info] ExtractFieldsReading.dijonParseAndThenToMap     100  thrpt    3    40391.142 ±   29894.312  ops/s
[info] ExtractFieldsReading.dijonParseAndThenToMap    1000  thrpt    3     3082.901 ±    3159.758  ops/s
[info] ExtractFieldsReading.dijonParseAndThenToMap   10000  thrpt    3      250.402 ±     157.485  ops/s
[info] ExtractFieldsReading.dijonParseAndThenToMap  100000  thrpt    3       18.290 ±      16.433  ops/s
```