@plokhotnyuk 2019-08-03
Hi, Pathikrit!

As you see the number of possible options to resolve the cache collision vulnerability is not so big:
1) change `toMap: Map` to `toMap: collection.Map` and use safe implementation like `java.util.LinkedHashMap`
2) introduce some configurable limit for the number of allowed keys in a map when calling `toMap: Map`
3) implement some simplified but safe version of `immutable.Map` for all supported Scala versions

IMHO the best option would be 3rd if it will be possible to write some version of immutable (radix-)tree map that is specialized for string keys and can fit the `Map[A, B]` interface.